statistics: move tests into GlobalStats test
