+++ statistics/BUILD.bazel
@@ -74,7 +74,7 @@ go_test(
     data = glob(["testdata/**"]),
     embed = [":statistics"],
     flaky = True,
-    shard_count = 39,
+    shard_count = 37,
     deps = [
         "//config",
         "//parser/ast",
