+++ statistics/handle/handletest/globalstats/BUILD.bazel
@@ -8,12 +8,13 @@ go_test(
         "main_test.go",
     ],
     flaky = True,
-    shard_count = 10,
+    shard_count = 12,
     deps = [
         "//parser/model",
         "//statistics/handle",
         "//testkit",
         "//testkit/testsetup",
+        "@com_github_pingcap_failpoint//:failpoint",
         "@com_github_stretchr_testify//require",
         "@org_uber_go_goleak//:goleak",
     ],
